Added correct version code and name.
This updates the version code and version name in the
`AndroidManifest.xml` to match the tagged release number.  I jumped
ahead a few versions to track the current eboard version (4.2.0),
but this might go out of sync at some point.

We need to keep this number up to date in the future to do android
market releases.