Added a failsafe behavior.
This failsafe behavior tracks the last-known point of connectivity
that the vehicle has had, testing if a particular address is reachable.

If the address becomes unreachable for longer than a timeout, the
vehicle begins navigating to the last reached point.

If it still cannot regain connectivity, it will begin navigating to the
preset home location.