Changed logging to directly print to file instead of using microlog. (#25)
Since we already route all vehicle logging through the VehicleLogger object,
using microlog does not really help us in any way.  This PR removes the
microlog4j dependency, and instead directly logs to a logfile directory.

This also changes logging to use a proper android directory request, so
logs will now be put into:
`/sdcard/Documents/platypus`

It also slightly changes formatting to put the full date at the top of
each logfile, instead of printing it on every line.  The timestamps used
in the log are relative to the start of the log, which makes the logs
much shorter.