Support for automatic usage and log uploads. (#29)
This PR adds automatic logging of server usage and log synchronization via Firebase.

All starts and stop of the server are logged to a local database which is synchronized automatically the next time that the application is run while on wifi. Each server has an entry containing version information and a "last update" timestamp.  The settings panel now has options to manually and automatically upload logfiles to the cloud.

Firebase authentication can be provided by means of a QR code containing the username and password of a valid Platypus vehicle login.  These QR codes should be secured similarly to other account information.