Attempt to create assisted teleop. Always run PID. Teleop adjusts desired angle.
Several large changes to VehicleServerImpl and AirboatController.POINT_AND_SHOOT.
Now there is a permanent TimerTask for updating the PID controller.
startWaypoints() only changes the waypoints queue.
POINT_AND_SHOOT now checks VehicleServerImpl.isAutonomous().
If autonomous, calculate desired heading as normal.
If not, it is because of teleop, so get desired angle from VehicleServerImpl.getDesiredAngle().
Note that the desired_angle variable is an AtomicInteger and represents desired angle in degrees!

POTENTIAL ISSUES:
Previously, the boat starts with _isAutonomous = false and no waypoints, and desiredAngle = 0.
Because the default thrust is going to be nonzero due it just being set to the P gain,
  the boat might immediately try to face whatever direction is 0 degrees. (East?)
Now, _isAutonomous starts = true. I don't know what issues that might cause right now.