LauncherFragment now sends sensor JSON on vehicle service start and on button press.
Ready for testing, but probably buggy.
Required binding the VehicleService to the LauncherFragment.
This unfortunately cluttered up the LauncherFragment considerably.
This needs to be tested, as my emulator seemed to run hot while the service was on.
I'm not sure if that means extra junk work is being done somehow, or if
  I just need to test it on a phone.