Added additional SensorData send commands for bluebox sensors. Added small delay before each SensorData message.
I am not sure what the root of the problem is, but the tablet is receiving duplicates when an ES2 is involved.
It is not consistent, but sometimes two temperature only SensorData objects arrive at the same time.
You always expect an EC and then a Temperature, but this is not the case.
Needs further debugging.